Move makeUseCases function
